# openapi.yaml
openapi: 3.0.3
info:
  title: RescueBite API
  description: |
    Food waste reduction platform. Restaurants list near-expiry food at a discount.
    
    **Key Features:**
    - Price decay state machine (Fresh → Discounted → Free → Compost)
    - Flash Auction — last 30 min before expiry, reverse auction (lowest bid wins)
    - Route Optimization — TSP-lite nearest neighbor algorithm for drivers
    - Geofencing — auto-complete delivery when driver enters 100m radius
    - Shelter role — charities receive free/compost food
  version: 2.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RouteOptimization:
      type: object
      properties:
        driverStart:
          type: object
          properties:
            latitude: { type: number }
            longitude: { type: number }
        totalStops: { type: integer }
        totalDistanceKm: { type: number }
        estimatedMinutes: { type: integer }
        algorithm: { type: string, example: "nearest-neighbor-tsp" }
        route:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [PICKUP, DELIVERY] }
              name: { type: string }
              address: { type: string }
              latitude: { type: number }
              longitude: { type: number }
              distanceFromPrev: { type: number, description: "km from previous stop" }
paths:
...
  /vendors:
    post:
      tags: [Vendors]
      summary: Create vendor profile
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [storeName, address, latitude, longitude]
              properties:
                storeName: { type: string, example: "Sushi Palace" }
                address: { type: string, example: "Almaty, Dostyk 123" }
                latitude: { type: number, example: 43.238 }
                longitude: { type: number, example: 76.945 }
      responses:
        '201': { description: Vendor created (pending approval) }
  /vendors/me:
    get:
      tags: [Vendors]
      summary: Get vendor profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Vendor profile }
    patch:
      tags: [Vendors]
      summary: Update vendor profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Updated }
  /vendors/me/orders:
    get:
      tags: [Vendors]
      summary: Get incoming orders
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Paginated orders }
...
  /drivers/me/route:
    get:
      tags: [Drivers]
      summary: "🗺️ ROUTE OPTIMIZATION — TSP-lite nearest neighbor algorithm"
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: orderIds
          required: true
          schema: { type: string }
          description: "Comma-separated order IDs, e.g. uuid1,uuid2,uuid3"
          example: "550e8400-e29b-41d4-a716-446655440000,550e8400-e29b-41d4-a716-446655440001"
      responses:
        '200':
          description: Optimized route with stops in order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RouteOptimization' }
        '400': { description: No orderIds provided or location not set }
...